+++ src/Console/TerminatableV63.php
@@ -4,12 +4,15 @@ declare(strict_types=1);
 
 namespace Faustoff\Contextify\Console;
 
-trait Terminatable
+trait TerminatableV63
 {
     use BaseTerminatable;
 
-    public function handleSignal(int $signal): void
+    public function handleSignal(int $signal): false|int
     {
         $this->processSignal($signal);
+
+        return false;
     }
 }
+